# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.4.0
  description: Korean Travel Guide API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /recommend-from-tags:
    post:
      operationId: recommendFromTags
      tags: [recommendations]
      summary: Guest preview using the GOAT score v2 engine
      deprecated: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecommendFromTagsRequest"
      responses:
        "200":
          description: Recommendations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecommendationsSuccessResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many recommendation requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /geocode-origin:
    post:
      operationId: geocodeOrigin
      tags: [recommendations]
      summary: Resolve an address or region query to an origin
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GeocodeOriginRequest" }
      responses:
        "200":
          description: Origin resolved
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GeocodeOriginSuccessResponse" }
        "404":
          description: Origin not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "429":
          description: Too many origin lookup requests
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /recommend-course:
    post:
      operationId: recommendCourse
      tags: [recommendations]
      summary: Build a day course from one selected GOAT recommendation card
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecommendCourseRequest"
      responses:
        "200":
          description: Day course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecommendCourseSuccessResponse"
        "400":
          description: Invalid request or selected place not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many recommendation requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /places/{id}:
    get:
      operationId: getPlace
      tags: [places]
      summary: Get a place by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: Place details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaceSuccessResponse"
        "404":
          description: Place not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: goat.sid
  schemas:
...
    ApiSuccessBase:
      type: object
      properties:
        success:
          type: boolean
          const: true
        code:
          type: string
        message:
          type: string
        requestId:
          type: string
      required: [success, code, message]
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          const: false
        code:
          type: string
        message:
          type: string
        data:
          nullable: true
        requestId:
          type: string
      required: [success, code, message, data]
...
    TravelPreferences:
      type: object
      properties:
        companion:
          type: string
          enum: [혼자, 연인, 친구, 가족]
        transport:
          type: string
          enum: [자차, 대중교통, 도보중심]
        visitTime:
          type: [string, "null"]
          enum: [새벽, 오전, 한낮, 오후, 일몰, 저녁, 야간, 밤/새벽, null]
        purpose:
          type: string
          enum: [가볍게 산책, 사진 위주, 액티비티, 조용한 휴식]
      required: [companion, transport, purpose]
      additionalProperties: false
    TravelOrigin:
      type: object
      properties:
        type:
          type: string
          enum: [current, region, address, skip]
        latitude:
          type: number
          minimum: -90
          maximum: 90
        longitude:
          type: number
          minimum: -180
          maximum: 180
        regionName:
          type: string
          minLength: 1
          maxLength: 200
      required: [type]
      additionalProperties: false
    RecommendFromTagsRequest:
      type: object
      properties:
        moodId:
          type: string
          minLength: 1
        referenceCardId:
          type: string
          minLength: 1
        travelPurpose:
          type: string
          enum:
            [
              사진·포토스팟,
              산책·힐링,
              카페·실내휴식,
              전시·건축관람,
              체험·액티비티,
              먹거리·야간탐방,
              숙소·리조트,
            ]
        transportType:
          type: string
          enum: [자차, 대중교통, 도보중심]
        visitTime:
          type: string
          enum: [새벽, 오전, 한낮, 오후, 저녁, 야간]
        currentMonth:
          type: integer
          minimum: 1
          maximum: 12
        debug:
          type: boolean
        preferences:
          $ref: "#/components/schemas/TravelPreferences"
        origin:
          $ref: "#/components/schemas/TravelOrigin"
        excludeIds:
          type: array
          uniqueItems: true
          maxItems: 61
          items:
            type: string
            minLength: 1
      additionalProperties: false
...
    GeocodeOriginRequest:
      type: object
      properties:
        query: { type: string, minLength: 2, maxLength: 120 }
      required: [query]
      additionalProperties: false
    GeocodeOriginData:
      type: object
      properties:
        origin: { $ref: "#/components/schemas/TravelOrigin" }
        address: { type: string }
        source: { type: string, enum: [ADDRESS, KEYWORD] }
      required: [origin, address, source]
      additionalProperties: false
    GeocodeOriginSuccessResponse:
      allOf:
        - $ref: "#/components/schemas/ApiSuccessBase"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/GeocodeOriginData" }
          required: [data]
    RecommendCourseRequest:
      type: object
      properties:
        recommendationId:
          type: string
          format: uuid
        selectedPlaceId:
          type: string
          minLength: 1
        primaryTheme:
          type: string
          enum:
            [
              바다·해안 무드,
              일본 소도시·골목 무드,
              알프스·고원·목장 무드,
              숲·정원·자연휴식 무드,
              레트로·시장·항구 무드,
              건축·전시·랜드마크 무드,
              휴양·카페·이국공간 무드,
            ]
        userMoodTags:
          type: array
          maxItems: 10
          items:
            type: string
            minLength: 1
        userSceneTags:
          type: array
          maxItems: 10
          items:
            type: string
            minLength: 1
        companionType:
          type: string
          enum: [혼자, 친구, 연인, 가족]
        travelPurpose:
          type: string
          enum:
            [
              사진·포토스팟,
              산책·힐링,
              카페·실내휴식,
              전시·건축관람,
              체험·액티비티,
              먹거리·야간탐방,
              숙소·리조트,
            ]
        transportType:
          type: string
          enum: [자차, 대중교통, 도보중심]
        radiusMeters:
          type: integer
          minimum: 100
          maximum: 20000
        maxCandidatesForLlm:
          type: integer
          minimum: 1
          maximum: 20
        forceRuleBasedFallback:
          type: boolean
        llmModel:
          type: string
          minLength: 1
          maxLength: 120
        debug:
          type: boolean
      required: [selectedPlaceId, primaryTheme]
      additionalProperties: false
...
    GoatRecommendationCard:
      type: object
      properties:
        rank: { type: integer, minimum: 1, maximum: 3 }
        role:
          type: string
          enum: [BEST_SCENE, SAME_MOOD_ALTERNATIVE, CONDITION_FIT_ALTERNATIVE]
        roleLabel: { type: string }
        placeId: { type: string }
        placeName: { type: string }
        city: { type: string }
        regionGroup: { type: string }
        primaryTheme: { type: string }
        placeType: { type: string }
        photoPoint: { type: string }
        recommendationUse: { type: string }
        bestTime: { type: string }
        seasonTags:
          type: array
          items: { type: string }
        purposeTags:
          type: array
          items: { type: string }
        accessibility:
          type: object
          properties:
            public_transport: { type: [string, "null"] }
            car: { type: [string, "null"] }
            walk: { type: [string, "null"] }
          additionalProperties: false
        note: { type: string }
        imageUrl: { type: [string, "null"] }
        address: { type: [string, "null"] }
        score: { $ref: "#/components/schemas/GoatScoreBreakdown" }
        reasons:
          type: array
          items: { type: string }
        cautions:
          type: array
          items: { type: string }
      required:
        [
          rank,
          role,
          roleLabel,
          placeId,
          placeName,
          city,
          regionGroup,
          primaryTheme,
          placeType,
          photoPoint,
          recommendationUse,
          bestTime,
          seasonTags,
          purposeTags,
          accessibility,
          score,
          reasons,
          cautions,
        ]
      additionalProperties: false
    Recommendation:
      type: object
      properties:
        place:
          $ref: "#/components/schemas/Place"
        role:
          type: string
          enum: [장면 최적, 같은 분위기 대안, 조건 맞춤]
        score:
          type: number
        reason:
          type: string
        matchedTags:
          type: array
          items:
            type: string
        scoreBreakdown:
          $ref: "#/components/schemas/LegacyRecommendationScoreBreakdown"
        safetyNotes:
          type: array
          items:
            type: string
        weatherFit:
          type: string
        parkingInfo:
          type: string
        routeInfo: { $ref: "#/components/schemas/RouteInfo" }
      required:
        [
          place,
          role,
          score,
          reason,
          matchedTags,
          scoreBreakdown,
          safetyNotes,
          weatherFit,
          parkingInfo,
        ]
    RecommendationsData:
      type: object
      properties:
        moodId:
          type: string
        referenceCardId:
          type: string
        appliedTags:
          type: array
          items:
            type: string
        seedPoolSize:
          type: integer
          const: 61
        candidatePoolSize:
          type: integer
          minimum: 0
          maximum: 61
        poolPolicy:
          type: string
          enum: [ALL61, PRIMARY43]
        poolReason:
          type: string
        fallbackUsed:
          type: boolean
        adaptivePoolRetryUsed:
          type: boolean
        recommendations:
          type: array
          minItems: 3
          maxItems: 3
          items:
            $ref: "#/components/schemas/Recommendation"
        cards:
          type: array
          minItems: 3
          maxItems: 3
          items:
            $ref: "#/components/schemas/GoatRecommendationCard"
        alternatives:
          type: array
          items:
            $ref: "#/components/schemas/GoatRecommendationCard"
        warnings:
          type: array
          items:
            type: string
        policyVersion: { type: string, enum: [goat-score-v2] }
        originStatus: { type: string, enum: [APPLIED, SKIPPED, UNAVAILABLE] }
        originNotice: { type: string }
        decisionAudit:
          type: object
          additionalProperties: true
      required:
        [
          moodId,
          appliedTags,
          seedPoolSize,
          candidatePoolSize,
          poolPolicy,
          poolReason,
          fallbackUsed,
          adaptivePoolRetryUsed,
          recommendations,
          policyVersion,
          originStatus,
          decisionAudit,
        ]
    RecommendationsSuccessResponse:
      allOf:
        - $ref: "#/components/schemas/ApiSuccessBase"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/RecommendationsData"
          required: [data]
...
    CourseStop:
      type: object
      properties:
        order: { type: integer }
        id: { type: string }
        title: { type: string }
        type:
          type: string
          enum: [START_PLACE, TOUR, CAFE, RESTAURANT, WALK, PHOTO, ETC]
        category:
          type: string
          enum: [START_PLACE, TOUR, CAFE, RESTAURANT, MARKET, WALK, PHOTO, ETC]
        address: { type: string }
        lat: { type: number }
        lng: { type: number }
        stayMinutes: { type: integer }
        reason: { type: string }
        source: { type: string }
      required: [order, id, title, type, category, stayMinutes, reason]
    StaticMapResult:
      type: object
      properties:
        provider:
          type: string
          enum: [KAKAO_JS_SDK_STATIC_MAP, KAKAO_MAP_SEARCH, NONE]
        staticMapConfig:
          type: object
          additionalProperties: true
        fallbackMapSearchUrl:
          type: string
        reason:
          type: string
      required: [provider]
    RecommendCourseData:
      type: object
      properties:
        status:
          type: string
          enum: [DONE, FAILED]
        resultType:
          type: string
          enum: [COURSE, UNKNOWN]
        mode:
          type: string
          enum: [LLM_OPENROUTER, RULE_BASED_FALLBACK]
        message:
          type: string
        selectedPlace:
          type: object
          additionalProperties: true
        conditions:
          type: object
          additionalProperties: true
        nearbyCandidateCount:
          type: integer
        courseTitle:
          type: string
        summary:
          type: string
        stops:
          type: array
          items:
            $ref: "#/components/schemas/CourseStop"
        staticMap:
          $ref: "#/components/schemas/StaticMapResult"
        llmPromptUsed:
          type: boolean
        failReason:
          type: [string, "null"]
        warnings:
          type: array
          items:
            type: string
      required:
        [
          status,
          resultType,
          mode,
          message,
          selectedPlace,
          conditions,
          nearbyCandidateCount,
          stops,
          staticMap,
          warnings,
        ]
    RecommendCourseSuccessResponse:
      allOf:
        - $ref: "#/components/schemas/ApiSuccessBase"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/RecommendCourseData"
          required: [data]
    OfficialTourInfo:
      type: object
      properties:
        contentId: { type: string, minLength: 1 }
        canonicalName: { type: string, minLength: 1 }
        address: { type: string }
        latitude: { type: number }
        longitude: { type: number }
        overview: { type: string }
        parking: { type: string }
        usageTime: { type: string }
        restDate: { type: string }
        phone: { type: string }
        homepage: { type: string }
        attribution: { $ref: "#/components/schemas/SourceAttribution" }
      required: [contentId, canonicalName, attribution]
...
    CurrentWeather:
      type: object
      properties:
        atKst: { type: string, minLength: 1 }
        sky: { type: string, enum: [CLEAR, CLOUDY, OVERCAST, UNKNOWN] }
        precipitation: { type: string, enum: [NONE, RAIN_OR_SNOW, UNKNOWN] }
        temperatureC: { type: [number, "null"] }
        windSpeedMps: { type: [number, "null"] }
        attribution: { $ref: "#/components/schemas/SourceAttribution" }
      required: [atKst, sky, precipitation, temperatureC, windSpeedMps, attribution]
    PlaceData:
      type: object
      properties:
        place:
          $ref: "#/components/schemas/Place"
        officialTourInfo:
          oneOf:
            - $ref: "#/components/schemas/OfficialTourInfo"
            - type: "null"
        externalPlaceInfo:
          oneOf:
            - $ref: "#/components/schemas/ExternalPlaceInfo"
            - type: "null"
        currentWeather:
          oneOf:
            - $ref: "#/components/schemas/CurrentWeather"
            - type: "null"
      required: [place, officialTourInfo, externalPlaceInfo, currentWeather]
    PlaceSuccessResponse:
      allOf:
        - $ref: "#/components/schemas/ApiSuccessBase"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/PlaceData"
          required: [data]